Refactor package structures
