+++ src/main/kotlin/io/memorix/messages/MessageDetailErrors.kt
