+++ src/main/kotlin/io/memorix/messages/MessageDetailSuccess.kt
