+++ src/main/kotlin/io/memorix/modules/databaseDI.kt
@@ -1,13 +1,10 @@
-package io.memorix
+package io.memorix.modules
 
 import io.memorix.database.DBConnector
 import io.memorix.database.DBConnectorFacade
 import org.koin.core.module.dsl.createdAtStart
 import org.koin.core.module.dsl.withOptions
-import org.koin.core.parameter.parametersOf
 import org.koin.dsl.module
-import io.ktor.server.routing.*
-import java.lang.System.getProperty
 
 val applicationDi = module {
     single<DBConnectorFacade> {
