+++ src/main/kotlin/io/memorix/modules/userDi.kt
@@ -1,5 +1,6 @@
-package io.memorix.user
+package io.memorix.modules
 
+import io.memorix.user.UserRepository
 import org.koin.dsl.module
 
 val userDi = module {
