+++ src/main/kotlin/io/memorix/routes/Routes.kt
@@ -1,4 +1,4 @@
-package io.memorix
+package io.memorix.routes
 
 object Routes {
     const val USERS_ROUTE = "/users"
